Chore : 멤버와 펫 패키지 분리 적용
